# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Performance Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  username: string;
}

export interface UserResponse {
  username: string;
  authenticated: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type SubjectStatsTopper = {
  name: string;
  score: number;
};

export interface SubjectStats {
  passedCount: number;
  failedCount: number;
  totalCount: number;
  passRate: number;
  failRate: number;
  averageScore: number;
  highestScore: number;
  lowestScore: number;
  topper: SubjectStatsTopper;
}

export interface StudentRecord {
  studentId?: string;
  studentName?: string;
  [key: string]: unknown;
}

export interface TopStudent {
  name: string;
  average: number;
  totalSubjects: number;
}

export interface Anomaly {
  type: string;
  subject?: string;
  description: string;
  count?: number;
  passRate?: number;
}

export type AnalysisResultSubjectWiseStats = { [key: string]: SubjectStats };

export interface AnalysisResult {
  totalStudents: number;
  totalSubjects: number;
  subjectWiseStats: AnalysisResultSubjectWiseStats;
  departmentPassRate: number;
  overallTopStudent?: TopStudent;
  topStudents: TopStudent[];
  studentsPassedAll: number;
  studentsFailedAny: number;
  averageScore: number;
  anomalies: Anomaly[];
  passPercentage: number;
  subjects: string[];
}

export interface ValidationResult {
  isValid: boolean;
  errors: string[];
  warnings: string[];
}

export interface AnalysisResponse {
  validation: ValidationResult;
  analysis?: AnalysisResult;
  rawData?: StudentRecord[];
  columns: string[];
  subjectColumns: string[];
}

export interface ExportRequest {
  analysisData: AnalysisResult;
  showStudentIds: boolean;
  rawData?: StudentRecord[];
  subjectColumns?: string[];
}

export type UploadFileBody = {
  file: Blob;
  /** Default pass percentage threshold */
  passPercentage?: number;
};