# lib/api-zod/src/generated/types/analysisResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Performance Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from "./analysisResult";
import type { StudentRecord } from "./studentRecord";
import type { ValidationResult } from "./validationResult";

export interface AnalysisResponse {
  validation: ValidationResult;
  analysis?: AnalysisResult;
  rawData?: StudentRecord[];
  columns: string[];
  subjectColumns: string[];
}