# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Performance Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultSubjectWiseStats } from "./analysisResultSubjectWiseStats";
import type { Anomaly } from "./anomaly";
import type { TopStudent } from "./topStudent";

export interface AnalysisResult {
  totalStudents: number;
  totalSubjects: number;
  subjectWiseStats: AnalysisResultSubjectWiseStats;
  departmentPassRate: number;
  overallTopStudent?: TopStudent;
  topStudents: TopStudent[];
  studentsPassedAll: number;
  studentsFailedAny: number;
  averageScore: number;
  anomalies: Anomaly[];
  passPercentage: number;
  subjects: string[];
}