# lib/api-zod/src/generated/types/anomaly.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Performance Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface Anomaly {
  type: string;
  subject?: string;
  description: string;
  count?: number;
  passRate?: number;
}