# lib/api-zod/src/generated/types/exportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Performance Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from "./analysisResult";
import type { StudentRecord } from "./studentRecord";

export interface ExportRequest {
  analysisData: AnalysisResult;
  showStudentIds: boolean;
  rawData?: StudentRecord[];
  subjectColumns?: string[];
}