# lib/api-zod/src/generated/types/studentRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Performance Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentRecord {
  studentId?: string;
  studentName?: string;
  [key: string]: unknown;
}